ints were not the problem

+++ Adafruit_SharpMem.cpp
@@ -272,9 +272,9 @@ uint8_t Adafruit_SharpMem::getPixel(uint16_t x, uint16_t y) {
     @brief Draw a fat line
 */
 /**************************************************************************/
-void Adafruit_SharpMem::drawFatLine(int x0, int y0, // first point
-                                    int x1, int y1, // second point
-                                    int strokeWidth,    // stroke width
+void Adafruit_SharpMem::drawFatLine(int16_t x0, int16_t y0, // first point
+                                    int16_t x1, int16_t y1, // second point
+                                    int16_t strokeWidth,    // stroke width
                                     uint16_t color) {
   if (strokeWidth < 1) {
     return;
